+++ src/main/java/minicraft/level/tile/Tile.java
@@ -37,7 +37,7 @@ public abstract class Tile {
 
     public final String name;
 
-    public byte id;
+    public short id;
 
     public boolean connectsToGrass = false;
     public boolean connectsToCloud = false;
